test: add test for sign pipeline
The unit tests needs a test key to work with. I have tried to create a
test keyring on the fly and while that worked I was not able to
successfully sign with that. gpg would bail with an ioctl error which I
didn't track down since using a static key works.