fix: Fix Brew formula class name when using custom name
The #597 added support for customizing the final Brew formula name.
The generated *.rb file still were using the project name as
class name and installing with custom name didn't work.
Fixed the issue by using the `name` field as an input for
building the class name, which defaults to project name
if no name given.