fix: Only put actual caveats in Homebrew formulae
If the config file has no entry for caveats, the Caveats field of the
struct defaults to the empty string. When strings.Split() is called on
the empty string, it returns []string{""}. This is considered truthy by
the template processor, resulting in a caveats section consisting of a
blank line.