fix: Use right architecture for ARM packages
Previously when creating deb package for ARMv6 and ARMv7 the
architecture information were invalid in the generated package.
For example ARMv6 package got armhf architecture which is not right.
According to https://wiki.debian.org/ArmHardFloatPort,
armel is armv4t, armv5, armv6 and armhf is armv7 so updated
to use correct architecture.