feat: Support per-format NFPM overrides
Support per-format NFPM overrides (as recently implemented in NFPM),
plus a pair or Goreleaser specific overrides, for "Replacements"
and "NameTemplate" (to address #640).

This keeps backward compatibility with existing config files:
settings at the root of the nfpm config section will still work as
previously (which is also useful to define common/shared settings
that we don't want to override anyway).