feat: Add the ability to place artifact name (#1338)
This commit exposes the Artifact name as a template variable so that the
artifact name can be moved around within the `targetURL`. This enable users to
manipulate the target URL as desired in order to pass in PUT parameters that
doesn't always following the currently defined `targetURL` which append the
artifact name to the end.

This is needed to address[1] by enable the ability to add metadata to
Artifactory REST API [2] as the URL parameters, which need to be after the
artifact names.

It's important that this is backward compatible with existing release
configurations so this is an opt-in option, if it's omitted or not set, no
changes to exist configurations. When enabled with `CustomArtifactName=True` as
part of HTTP Upload options `Artifact.Name` will no longer be appended to the
end of TargetURL.

[1]:https://github.com/goreleaser/goreleaser/issues/1336
[2]:https://www.jfrog.com/confluence/display/RTF/Artifactory+REST+API#ArtifactoryRESTAPI-Example-DeployinganArtifact