feat: allow NFPM packages to be used in Docker builds
This allows packages built by the NFPM pipe to be accessed from within
Docker builds. The NFPMs to expose are listed in a new `nfpms` key in
the Docker config, defaulting to the first NFPM if there is only one
Docker section.

The new code mostly parallels the existing code for retrieving binaries.
To keep things simple, all built packages for a given NFPM ID are made
available.