feat: run hooks as last step in the release pipeline
Hello my goreleaser friends!

!!! I would like to make this PR better (will explain what I mean
later), it is here to gather feedback about the feature itself.

I asked about running some custom code at the end of the release cycle
on Gopher Slack today.

I would like to manipulate the manifest for a bunch of Docker
images I create during the release life cycle.

My build generates docker images for multiple architectures
and in order to teach the registry about that you are supposed to
generate or updated a manifest.

You have to do it when the images are pushed to the registry.

It looks like a custom publisher won't work because it runs before
the docker one.

This PR proposes the same approach we use for the `before` feature, but
it runs at the end of the release workflow, as last step.

---

As I said, I just copy pasted and hooked a bunch of code. If this has
sense, I think I would like to write a pipe called `HookRunner` in order to reuse
it for Before and After.

---

- [ ] Add documentation
- [ ] Add tests