feat: IDs filter for Checksums pipe (#1805)
* Add IDs filter to checksums pipe

This allows specific artifacts to be excluded from the checksums file.

The reason for introducing this, is the requirement of Terraform
Registry releases to only contain the prescribed archives in the
checksums file. If the file contains more, the release is not accepted
by the Terraform Registry.

* Add test case for IDs filter of checksums pipe

* Document IDs filter of checksum pipe

* Always apply type filter for artifacts

This is more in line with how the other ID-filters work.