fix: Enable darwin/arm64 builds for go1.17 and later
This refactors the check for disabling darwin/arm64 builds on Go versions less than
go1.16. Previously, it would only match on the substring 'go1.16', which limits this to
only work with go1.16 (not anything later).

The new isGo116orLater() function now verifies the go version matches based on a regex,
due to the slightly unconventional style of go version strings. The regex pulls out the
version digits, and verifies that the major version is >=1 and the minor version is >=16
(if major version == 1).

The isGo116orLater() function also now runs the GoBinary command from the 'build.Dir',
which resolves issues with relative 'gobinary' paths.