feat: release only specific build ids
In order to allow releasing specific builds I added a new argument to release called 'only-build'. This argument allows one to select which build ids to build by using a regular expression. This is handy when you only want to release a specific build or test a specific build without having to wait for everything else to build as well.