fix: nfpm globbing multiple times
The nfpm pipe was globbing the input list, and then globbing its result
as well. If you add a folder that had a file with, say `[]`, in it, it
would later fail the release upon trying to evaluate that as a glob.

This is probably some faulty nFPM update in which I should have removed
that but didn't.

Either way, this should fix it.

closes #2946