fix(nix): sourceRoot when using archives.wrap_in_directory
If `archives.[*].wrap_in_directory` is set, it'll create a folder inside
the archive file, usually something like `app_goos_goarch`.

In those cases, the root of the archive is not constant, so we create a
`sourceRootMap` and use it instead.

In cases where the `sourceRoot` is constant, the generated derivation
will be the same.

Signed-off-by: Carlos Alexandro Becker <caarlos0@users.noreply.github.com>