fix: snapcraft temporary directory + concurrency
this bug comes and goes every couple of versions it seems.

this will change the snapcraft implementation to run the first item
without concurrency, so all needed shared directories can be created
without issues, and then grows the limit of the wait group so the other
ones can run in parallel.

I haven't tested this yet, but I think it'll work.

- [ ] test
- [ ] godoc

refs https://github.com/goreleaser/goreleaser/issues/1715
refs https://bugs.launchpad.net/snapcraft/+bug/1889741