fix: signature template failed silently after signing process completed
The presence of an artifact field in the
`signature` or `certificate` template field caused
a silent failure in the template when re-applied
after the external signing process was called.

This was due to the artifact being presence in the
template context before the signing process, but
not after. An error here was also ignored.

The fix supplies the artifact to the template
context, and also allows a template failure to
fail the overall process.