fix(changelog): fix random order of first commit in first release
This is a bug that occurs randomly under a very specific condition.

In the `TestGroup` unit test, we are making 11 git commits, all within a
second. There seems to be a bug with `git log` where it is unable to
order the commits in reverse chronological order if all commits have the
same authored and committed date, as shown below.

`git log` without revision range:

	/tmp/TestGroup4125952855/001 main ❯ git log --oneline
	85f005f (HEAD -> main, tag: v0.0.2) this is not a Merge pull request
	27dbd0e bug: Merge pull request  from goreleaser/some-branch
	3495034 feat: added that thing
	9f0db77 chore: something about cArs we dont need
	634e043 docs: whatever
	ef52fef fix: whatever
	ff49bea feat(deps): update foobar [bot]
	e0f4e4b ignored: whatever
	ce7fbfa fixed bug 2
	940a684 added feature 1
	2750980 (tag: v0.0.1) first

`git log` with multiple revision ranges, the "first" commit appears at the top:

	/tmp/TestGroup4125952855/001 main ❯ git log --oneline 2750980 v0.0.2
	2750980 (tag: v0.0.1) first
	85f005f (HEAD -> main, tag: v0.0.2) this is not a Merge pull request
	27dbd0e bug: Merge pull request  from goreleaser/some-branch
	3495034 feat: added that thing
	9f0db77 chore: something about cArs we dont need
	634e043 docs: whatever
	ef52fef fix: whatever
	ff49bea feat(deps): update foobar [bot]
	e0f4e4b ignored: whatever
	ce7fbfa fixed bug 2
	940a684 added feature 1

If we specify only one revision, then the commits are ordered correctly
in reverse chronological order:

	/tmp/TestGroup4125952855/001 main ❯ git log --oneline v0.0.2
	85f005f (HEAD -> main, tag: v0.0.2) this is not a Merge pull request
	27dbd0e bug: Merge pull request  from goreleaser/some-branch
	3495034 feat: added that thing
	9f0db77 chore: something about cArs we dont need
	634e043 docs: whatever
	ef52fef fix: whatever
	ff49bea feat(deps): update foobar [bot]
	e0f4e4b ignored: whatever
	ce7fbfa fixed bug 2
	940a684 added feature 1
	2750980 (tag: v0.0.1) first

This bug can only happen when all commits are created at the same time,
and the user is creating the first release note.

This commit fixes the bug by excluding the first commit SHA-1 hash from
`git log` in `gitChangeLogger`.

Signed-off-by: Eng Zer Jun <engzerjun@gmail.com>