fix: group by platform when using rust or zig mixed with go
zig and rust builders don't set goamd64 and others in the artifacts, so
places that use GroupByPlatform (e.g. nFPM) will have two groups, even
if you're using the default goamd64 in the go builds.

this pr fixes that, trying to use the basic goos+goarch to group (as all
builders set it), and potentially disambiguating if more than 1 goamd64
is set.

same deal with gomips and goarm.

closes https://github.com/goreleaser/goreleaser/issues/5371