fix(archive): make sure binaries are executable (#5631)
On windows, files have no executable bits.

We read the file permissions on disk to decide its permissions inside
the archives - this means that, on windows, the binaries will not have
the executable bit when unextracted.

This sets the build file info mode to 0o755 by default, which should
resolve this issue.

This was also fixed at some point in time in the past, but apparently
regressed. Now that we test on windows as well, this shouldn't happen
again.

closes #5632

Signed-off-by: Carlos Alexandro Becker <caarlos0@users.noreply.github.com>