fix(log): custom git log format and parsing (#5635)
JSON formatting was brittle, as quotes et al were note escaped. Thought
about doing it with xml.Unmarshal, but it would have the same issue, but
this time with unmatched `<>`.

This uses a custom XML-based format, but does not use `encoding/xml` to
parse it.

thanks @raphamorim for reporting this (in Discord)