fix(nix): use nix-hash instead of nix-prefetch-url (#5669)
`nix-prefetch-url` will need the release to be published to work.

If the user wants to start the release as a draft and publish manually,
this will fail, as the URL will fail to resolve.

Using `nix-hash` instead is also faster than `nix-prefetch-url`, as it
doesn't need to download the artifact again.