fix(nix): snapshot builds can now hash files as well (#5894)
Historically, we used to use `nix-prefetch-url`, so it could only
calculate the hashes after the artifacts were pushed, so when using
`--snapshot`, hashes would always be `0000...`.

Rather recently, we've changed it to use `nix-hash` instead, so we could
also have changed it so it uses real hashes all the time.

This PR does that.

Discussion: https://github.com/orgs/goreleaser/discussions/5893

Signed-off-by: Carlos Alexandro Becker <caarlos0@users.noreply.github.com>