fix(sbom): --enrich=all should be the default (#6095)
currently the generated SBOMs might not be too useful without enriching,
as they won't have dependencies licenses and things like that.

enriching fixes it. It's done on goreleaser `main`, not sure if we
should make this the default or not 🤔

you can verify a SBOM with:

```sh
jq -r '.components[] | .name + " " + ([.licenses[]?.license.id] | join(","))' file.sbom.json
```

Signed-off-by: Carlos Alexandro Becker <caarlos0@users.noreply.github.com>