feat: add Discourse announcer (#6199)
This PR adds a new Announcer for
[Discourse](https://discourse.org/)-based forums. This is a bare minimum
PR. There's more I could do with error codes and with allowing
threading. I will wait-and-see for those and if there's interest, I can
add that in a future PR.

I will comment in-line on why I made certain decisions and/or ask
questions.

Closes: #6196

---------

Co-authored-by: Carlos Alexandro Becker <caarlos0@users.noreply.github.com>