feat: support ko loading base image from daemon (#6233)
**What**
Allow users to use a locally cached image as the base image for ko. 

**Why**
The current default behaviour is trying to fetch the image remotely, but
this is not always the expected behaviour. E.g ko will support this
behaviour through `defaultBaseImage`

Note: this could be hidden behind a flag, and this may be a better
approach from what this PR does which is to prioritize local images.

---------

Co-authored-by: Carlos Alexandro Becker <caarlos0@users.noreply.github.com>