fix(github): improve create/update file (#6280)
I've seen that since today GitHub API seems to be failing when creating
new files through this endpoint, I think they changed something in their
implementation, and, technically, I think we were sending an invalid
request (empty SHA instead of `nil` SHA).

If the file don't yet exist, we should leave the SHA `nil` instead. 
If that's the reason, this should fix it.

Signed-off-by: Carlos Alexandro Becker <caarlos0@users.noreply.github.com>